Now scheme=2 works with humptest.
Fixed usage of empty strings: not allowed in fortran, changed
them to ' '

Fixed calls of readkey, with third parameter ' '. This is an
output parameter of readkey.

Fixed calls to min and max, with different kinds of parameter
(one double, the other single precision)

Adapted makefile, now, when USEXLF is set, compilation is done
using the IBM xlf90 and mpfort compilers.

Adapted makefile: do not use mpi-compiler for compilation of 
makeincludes to avoid problems on IBM linux system.
 


git-svn-id: https://svn.oss.deltares.nl/repos/xbeach/trunk@51 86136381-33bf-4961-916e-47ea43f679af